refactor: colocate components in feature directories
