--- .mcp.json
@@ -1,8 +1,0 @@
-{
-  "mcpServers": {
-    "browsermcp": {
-      "command": "npx",
-      "args": ["@browsermcp/mcp@latest"]
-    }
-  }
-}
