+++ package.json
@@ -13,8 +13,8 @@
     "test:watch": "vitest --watch",
     "test:ui": "vitest --ui",
     "test:coverage": "vitest --coverage",
-    "test:e2e": "playwright test",
-    "test:e2e:ci": "playwright test --reporter=list",
+    "test:e2e": "playwright test --reporter=list",
+    "test:e2e:browser": "playwright test",
     "test:e2e:ui": "playwright test --ui",
     "test:e2e:debug": "playwright test --debug",
     "test:e2e:headed": "playwright test --headed",
