+++ src/features/market/components/MarketSection.tsx
