+++ src/ui/button/button.tsx
@@ -1,4 +1,4 @@
-import Spinner from '@/components/ui/spinner';
+import Spinner from '@/ui/spinner';
 
 export interface ButtonProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement> {
