+++ src/ui/spinner/Spinner.module.scss
